backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
pydantic-settings==2.8.1
python-multipart==0.0.20
pytest==8.3.5
httpx==0.28.1
langchain==0.3.27
langchain-openai==0.3.29
sqlalchemy==2.0.39
pymysql==1.1.1
cryptography==44.0.2
python-docx==1.1.2
pypdf==5.4.0
pillow==11.2.1
pytesseract==0.3.13
